Refactor LocationHelper
- Rename it to LocationRepository
- Move it to data.repository package
- Make it Singleton
